Fixed deployment issue with cloudflare pages
It seems that cloudflare pages needed a different
version of node 22 to work, which it asked me to
set in my asdf's `.tool-versions` file.

I have also added wranger which is cloudflare's
cli tool as a dependency to the project and added
a preview script, even though I do not yet fully
understand the deployment process on this environ-
ment and what preview means in this context.

Let's see!