Fixed the issue with 'Hello world'
As it turns out, Cloudflare has two types of projects
which can be created in a similar fashion:
- worker projects
- pages projects

They are both accessed from the same part of the
website but offer differences in configuration,
and behavior.

I think that what I did wrong earlier was creating
a worker project that had a token which was then
modified to allow the deployment to upload static
files, which were simply being ignored by the worker.

I am assuming that now, once I create the write type
of project, that things will work effortlessly,
let's see!