Switched from pnpm to npm to make cloudflare work
The deployments at cloudflare are still failing,
this time around because it is unable to use the
asdf tool to install the right version of pnpm
which is dictated in the .tool-versions file

I am hence switching from pnpm to npm (with 300MB
at least of additional overhead) to see if it would
work.

Incredibly, incredibly painful process.