Added PGP Public Key copy functionality
In addition to adding the public key to the public
folder, I also added the public folder to the special
include paths using `@public`, which allows me to
import it as raw text using:
```typescript
import PGPPublicKeyText from '@public/public.pgp?raw'
```

Additionally, I created a hidden HTML container to
store the statically-imported icons for the source
code, PGP public key and the print functionality,
which were previously hard coded.

I also renamed Print.astro in icons, which was named
in a misleading manner, to Shortcut.astro, which
more clearly designates its function as the icon
to be shown at narrower screen widths to open up
the keypad menu.