Ensured that page breaks do not occur mid content
In print layout, it used to happen that headings
were stranded on one page with the following content
being pushed to the next page. This was fixed using
the `page-break-after: avoid` CSS directive.

Additionally, I added a class `.print-together` to
ensure that experience and project entries are not
split into two parts due to page breaks which looked
untidy.