Improved CV content, fixed language, split skills
The CV content had issues, namely with the job
highlights being named something else (key_resp.)
following the schema of another project which I
migrated my data from using poor AI.

Additionally, I split the skills into core and tech
skills while also adding expert skills, to be shown
differently (darker).

The core skills have no icons and are capitalized
programmatically.

Finally, a persistent bug that should not have made
it as long as it did, is the language of the page
being set to Spanish (remnant from the original
template), which explained why Chrome kept on re-
commending to translate the page from Spanish.

This was also fixed and the page has its language
now set to English in the <html> tag.