Fixed hyperlink bg and added 'samples' hidden path
The hyperlink icon was previously placed in the
public folder which made it un-optimisable by astro

I got around this by using an import with a `?url`
vite option, which resulted in importing a data url
for some reason, then I was faced with the challenge
of getting this into the public styles. Apparently
it is not possible to interpolate js variables into
CSS directly so I had to use a feature in astro that
translate them into CSS variables. This created
another complication since it was really difficult
to use the variable as a path to the background-
image, as it turns out, there is a workaround which
is explained in question 42330075 on stackoverflow
which involves introducing `url("${url}")` as the
variable content. This finally worked, which feels
like it had taken much more of my time than it needed
to. Anyhowsss

Second I improved the content, and I also added a
hidden path '/samples' which I needed to print an
additional supplementary page to my PDF CV to attach
to a job application as a "Sonstige Unterlagen" or
a special document. I also added a new project and
enhanced the content a bit.

One more thing, I got some feedback on the color of
the less mastered skills hence I changed the color
from grey to something else.